backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "fast-flights>=2.2",
    "firecrawl-py>=3.2.1",
    "langchain>=0.3.27",
    "langchain-exa>=0.3.0",
    "langchain-openai>=0.3.30",
    "langgraph>=0.6.5",
    "loguru>=0.7.3",
]
